texture: Add support for mipmapping
Each texture's pixels are now given as a series of mipmaps, where level
0 is the base texture, and levels above that are successively smaller
versions of the base image.

Mipmapping is not yet enabled for actual rendering, though - this commit
only modifies the way texture data is loaded in and stored prior to it
being uploaded to the renderer. Future commits will add mipmapping to the
render pipeline.