Remove win32 suffix from file and function names
(Except from window_win32.c/.h, for clarity.)

For the moment, the program targets the Win32 platform exclusively, so
there's no need at the moment to specify which files and functions are
Win32-specific.

Though it's nice to have that distinction anyway as a way of future-
proofing, I think having the suffix in these names makes them a bit too
cumbersome to read. Should the program be expanded onto other platforms
in the future, the question of namespacing can be brought up again then.