surface/direct3d-7: Fix crash on surface release
Fixes what was a crash with a page fault in Wine on releasing the
renderer surface.

I'm not totally sure whether this was just a Wine thing, and I have a
feeling there may be some memory issues going on, but I don't have
a memory debugger on hand. Let's just assume for now that the issue
was real and is now fixed - but later testing will verify.